Use LVM snapshots to speed up instance creation
When use_cow_images is set, speed up the creation of instances by
caching the base image as a logical volume and then creating snapshots
of that for use by the instances.

Known issue: stale logical volume caches of base images aren't cleared
automatically.

Note: this is very much a work in progress.  For now, it was written
with minimizing impact on the code base, and as of yet has no tests.

DocImpact: since use_cow_images is set by default, this changes the default
behavior of the LVM backend.