Use LVM snapshots to speed up instance creation
When use_cow_images is set, speed up the creation of instances by
caching the base image as a logical volume and then creating snapshots
of that for use by the instances.

If the appropriate configuration variables are set (libvirt_images_type
= 'lvm' and use_cow_images = True), lvm caches are cleared periodically
if not in use, together with their respective base images.

DocImpact: since use_cow_images is set by default, this changes the default
behavior of the LVM backend.